Add decorator to keep session id
The _keep_id decorator helps to switch from browser windows without loosing the user session id. This decorator is using the logic that was part of the access_feed method.
For the specific case of the login function the new window must be opened only after the sign in process.
This commit also include changes in the conditions of the loop in the get_followers method.